HADOOP-2168. Fix problem with C++ record reader's progress not being reported to framework. (acmurthy via omalley)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@689635 13f79535-47bb-0310-9956-ffa450edef68

+++ src/examples/pipes/impl/wordcount-nopipe.cc
@@ -105,7 +105,7 @@ public:
    */
   virtual float getProgress() {
     if (bytesTotal > 0) {
-      return bytesRead / bytesTotal;
+      return (float)bytesRead / bytesTotal;
     } else {
       return 1.0f;
     }
